refactor: extract ExportRetryConfiguration.swift from TrackerExportRetryTypes.swift
Splits the `ExportRetryConfiguration` Sendable struct (with immediate
static preset) out of TrackerExportRetryTypes.swift.

Byte-identical move of lines 3-15. Both files import Foundation.

TrackerExportRetryTypes.swift: 37 → 23 lines. Tests: 667.
